Refactor into components and modernize styles
