+++ src/App.js
@@ -1,155 +1,5 @@
-import {useState} from 'react';
+import Board from './components/Board';
 
-function Square({value, onSquareClick, highlight}) {
-  return (
-    <button
-      className={`square${highlight ? ' highlight' : ''}`}
-      onClick={onSquareClick}
-    >
-      {value}
-    </button>
-  );
+export default function App() {
+  return <Board />;
 }
-
-export default function Board() {
-  const [xIsNext, setXIsNext] = useState(true);
-  const [squares, setSquares] = useState(Array(9).fill(null));
-
-  function handleClick(i) {
-    if (squares[i] || calculateWinner(squares)) {
-      return;
-    }
-    const nextSquares = squares.slice();
-    if (xIsNext) {
-      nextSquares[i] = 'X';
-    } else {
-      nextSquares[i] = 'O';
-    }
-    setSquares(nextSquares);
-    setXIsNext(!xIsNext);
-  }
-
-  const winnerInfo = calculateWinner(squares);
-  const winner = winnerInfo ? winnerInfo.winner : null;
-  let status;
-  if (winner) {
-    status = 'Winner: ' + winner;
-  } else if (squares.every((square) => square !== null)) {
-    status = 'Draw';
-  } else {
-    status = 'Next player: ' + (xIsNext ? 'X' : 'O');
-  }
-
-  function resetGame() {
-    setSquares(Array(9).fill(null));
-    setXIsNext(true);
-  }
-
-  const winningLine = winnerInfo ? winnerInfo.line : [];
-
-  return (
-    <>
-      <div className="status">{status}</div>
-      <div className="board">
-        <div className="board-row">
-          <Square
-            value={squares[0]}
-            onSquareClick={() => handleClick(0)}
-            highlight={winningLine.includes(0)}
-          />
-        <Square
-          value={squares[1]}
-          onSquareClick={() => handleClick(1)}
-          highlight={winningLine.includes(1)}
-        />
-        <Square
-          value={squares[2]}
-          onSquareClick={() => handleClick(2)}
-          highlight={winningLine.includes(2)}
-        />
-      </div>
-        <div className="board-row">
-          <Square
-            value={squares[3]}
-            onSquareClick={() => handleClick(3)}
-            highlight={winningLine.includes(3)}
-          />
-        <Square
-          value={squares[4]}
-          onSquareClick={() => handleClick(4)}
-          highlight={winningLine.includes(4)}
-        />
-        <Square
-          value={squares[5]}
-          onSquareClick={() => handleClick(5)}
-          highlight={winningLine.includes(5)}
-        />
-      </div>
-        <div className="board-row">
-          <Square
-            value={squares[6]}
-            onSquareClick={() => handleClick(6)}
-            highlight={winningLine.includes(6)}
-          />
-        <Square
-          value={squares[7]}
-          onSquareClick={() => handleClick(7)}
-          highlight={winningLine.includes(7)}
-        />
-        <Square
-          value={squares[8]}
-          onSquareClick={() => handleClick(8)}
-          highlight={winningLine.includes(8)}
-        />
-        </div>
-      </div>
-      <button className="reset" onClick={resetGame}>
-        Reset Game
-      </button>
-    </>
-  );
-}
-
-function calculateWinner(squares) {
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6],
-  ];
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return {winner: squares[a], line: lines[i]};
-    }
-  }
-  return null;
-}
-
-// a component is a piece of reusable code that represents a part of a user interface
-
-// <button> is a JSX element, a combination of JS code and HTML tags; className = "square" is a property of the <button> element
-
-// use Fragments (<> and </>) to WRAP multiple adjacent JSX elements
-
-/* "lifting state up" refers to moving the shared state to the closest common
-  ancestor of components that need to share that data.
-
-  When multiple components need access to the same state, instead of duplicating it in each component, you move the state to their parent component and pass it down via props. This allows the parent to control the state and ensures consistent data across all child components.
-
-  Steps:
-  1. Move the state to the common parent.
-  2. Pass the state and state updater function as props to child components.
-  3. Children use these props to render and update the shared state. */
-
-// JavaScript supports closures which means an inner function (e.g. handleClick) has access to variables and functions defined in a outer function (e.g. Board). The handleClick function can read the squares state and call the setSquares method because they are both defined inside of the Board function.
-
-// A closure is a function that remembers the variables from the function in which it was created
-
-// onClick is special to React in that it automatically attatches an event listener, this doesnt mean you can pass a function as an argument, you must pass a reference to a function
-
-// It does not matter whether you define calculateWinner before or after the Board. Let’s put it at the end so that you don’t have to scroll past it every time you edit your components.
