+++ src/styles.css
@@ -3,97 +3,69 @@
 }
 
 body {
-  font-family: sans-serif;
   margin: 0;
   padding: 0;
+  font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif;
+  background: #f5f5f5;
   display: flex;
-  flex-direction: column;
   align-items: center;
   justify-content: center;
-  height: 100vh;
+  min-height: 100vh;
 }
 
-h1 {
-  margin-top: 0;
-  font-size: 22px;
+.game-container {
+  text-align: center;
 }
 
-h2 {
-  margin-top: 0;
-  font-size: 20px;
+.status {
+  font-size: 1.5rem;
+  font-weight: 600;
+  margin-bottom: 1rem;
 }
 
-h3 {
-  margin-top: 0;
-  font-size: 18px;
-}
-
-h4 {
-  margin-top: 0;
-  font-size: 16px;
-}
-
-h5 {
-  margin-top: 0;
-  font-size: 14px;
-}
-
-h6 {
-  margin-top: 0;
-  font-size: 12px;
-}
-
-code {
-  font-size: 1.2em;
-}
-
-ul {
-  padding-inline-start: 20px;
+.board {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 8px;
+  width: min(90vmin, 300px);
+  margin: 0 auto;
 }
 
 .square {
   background: #fff;
-  border: 1px solid #000000;
-  float: left;
-  font-size: 24px;
+  border: 2px solid #333;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 2rem;
   font-weight: bold;
-  line-height: 34px;
-  height: 50px;
-  margin-right: -1px;
-  margin-top: -1px;
-  padding: 0;
-  text-align: center;
-  width: 50px;
+  aspect-ratio: 1 / 1;
+  cursor: pointer;
 }
 
 .square.highlight {
-  background: #90ee90;
+  background: #e0ffe0;
 }
 
 .reset {
-  margin-top: 15px;
-  padding: 8px 12px;
-  font-size: 16px;
+  margin-top: 1.5rem;
+  padding: 0.5rem 1rem;
+  font-size: 1rem;
+  border: none;
+  background: #333;
+  color: #fff;
+  cursor: pointer;
 }
 
-.board-row:after {
-  clear: both;
-  content: '';
-  display: table;
+.reset:hover {
+  background: #555;
 }
 
-.status {
-  margin-bottom: 10px;
-}
-
-.board {
-  display: inline-block;
-}
-.game {
-  display: flex;
-  flex-direction: row;
-}
-
-.game-info {
-  margin-left: 20px;
+@media (min-width: 600px) {
+  .square {
+    font-size: 2.5rem;
+  }
+  .board {
+    width: 300px;
+  }
 }
